General: Fix paired palettes (GetDlgCtrlID issue)
In the 2008 1.22 release, some versions would crash when hovering over a palette item due to a misuse of GetDlgCtrlID.  In the 1.23 initial release I made a QnD fix to disable that usage, figuring I would come back to when it mattered.  It turns out that that fix disabled editing paired palettes (Captain America and his shield, Captain Commando's ninja gang, etc).  Given that that code actually matters, I went back and made the correct for for the GetDlgCtrlID crash.