Unicode support
This is actually a workaround for a bug in MFC.  The MfcLink control corrupts memory when dealing with MBCS strings.  So when I added a simple hyperlink to the About page... PalMod started crashing.  I have reported the bug to MSFT.  In the interim, this change makes the painful but arguably long overdue switch to build Unicode.

I kept the file readers (PNG, extras) ANSI as required, and ran through a pretty solid battery of usage fixing up a couple hiccups.  However this is one of the biggest changes to date, so I'll keep aneye on usage reports.