MVC2-Arcade: Adjust hexes
Our MVC2-A hex list was sourced from first color, not transparency color.  We accounted for that in most usage in code through simply walking forward one when reading/writing.  But that runs afoul of the proc_supp code which is operating on the exact buffer locations.  So I instead did the right thing and corrected the ... all the hexes we had.

Also improved Rogue and Jin's proc_supp logic, added descriptions for Rogue, and fixes Bonerine's claws offset.