UI spinner; work around linker bug; updated palette hexes
Added UI spinners for ARGB tweaking.  Users said it would be easier to work with.

The LTCG:incremental option is causing the "Unknown NEO*GEO" and KOF2002 code to collide and crash: turned that off for now.  Appears to be a new bug in the latest MSVC linker.

Updated the hexes for Jojos, KOF202UM, and MvC2.