SFIII3_A: Move to table-based layout; Jojos: rename Load Dir entries
SFIII3 is the important change here.  We had been using the old calculation based layout from 2007, but as we discover new palettes that had made it difficult to inject those excellently.  So this change moves SFIII3 to the 2020 table-based layout, as that makes it trivial to add in and update palettes as needed.