Code cleanup
Moved a bunch of common game loading code to CGameGlass.  We can eliminate the InitDataBuffer overrides in the future if we handle SelectedROM in CGameClass: that's on my roadmap.  Only MVC2 is currently using the UpdatePalData base implementation: proliferating and tweaking that is also on the roadmap.

Also fixed MVC2-A not reading alpha files on game load.