ST, 3S, SVC, fix signed/unsigned comparison
ST: Remaining extra and super trail palettes.

3S: More stages.

SVC: Don't show the 'encrypted' warning for SVCSPlus since it's also the same ROM set for SVCSPlusA, which is decrypted.

We were decrementing an UINT, which worked great when everything was of an expected length.  But Blanka has a few 15 color palettes ... and thus we move decrementing counters to INT.