IPS support; 3S: Alex join
This changelist adds support for International Patching System (IPS) patch generation for games we support one ROM for.  We will create a that patch file using all changes generated in the current session: we won't have data for previous sessions.  We do not currently support patch generation for games we use multiple files to support: that logic simply hasn't been built yet, as the Rules logic for DIR and SIMM based games needs to be further abstracted.