MvC2-A: more logical alpha support
Previously you could have a modified alpha build but only be able to see the values after you turned on Allow Transparency and reloaded the game.  Now we always will store the actual values, and the only delta when clicking Allow Transparency is that we will enable the A color controls.

Minor code cleanup on Karnov/ImgDump.  Originally I thought ImgDump was handling alpha incorrectly for MvC2-A, but then noticed that it was because MvC2-A was ignoring the true alpha values when we recreated the palettes for imgdump.