Consolidate palette updates; custom alpha mode
There were a number of forks of UpdatePalData to handle the variable needs games have.  This checkin consolidates all of those down to one base implementation shared by everything except the ancient MvC2 code.  Syntax is pretty easy: just set the CreatePalOptions and the Alpha Mode and the revised UpdatePalData handles it all for you.

As part of this, we now expose current Alpha mode in the menus.  This is RO except if you are using Unknown Game mode, where you can customize it as needed for your game.