KOF99AE support;improved Fighter Factory support;MvC2 and KOF03 tweaks
When we flip Fighter Factory palettes now, we only apply the known useful portions if we're going to be splitting across multiple palettes.  So if we enter in the unused black region of the file for palette 2, we just leave the old palette 2 info.

MvC2: show the core character sprite for various largely unused effect palettes.

KOF03: new effect palettes