Secondary palette processing for COTA
We had a system for MvC2 wherein we would update secondary/connected palettes to help prevent users having bad flashing effects as the character reverted to an unchanged palette for special effects.  I've now adapted this code for more generalized usage and plugged it in for COTA as a starting point.

Also fixed the lookup for Felicia's Zapped palette.