palette updates; file/palette UINT16s -> size_t; vector some arrays
There's a handful of palette updates here, but the bulk of the change is code cleanup.  This change updates all usage of files/palette counts to use size_t as well as promoting some array usage to std::vector.  This should be a no-op that just enables simplifying a few code paths that were bothering me.