mvc2-dc-16: auto multisprite export for Extra color-specific palettes
One of the advantages of the previous implementation of MVC2::PalSel was that it baked in the knowledge needed for multisprite export.  In this new world order, that information is now derived from the move description set.  We use button label key names to determine whether or not something is eligible for pairing, and if that's true then we determine pair distance and logic from there.