Simplify Extras load, code cleanup
The Extra code is tailored to allow insertion of extras both as an dedicated Extra Palettes unit and to append extras to character units.  This was used only for SF3 arcade (Oro's Turtle/Tengu, Urien's Intro) in PalMod2007 but has been made obsolete as I've restructured PalMod to directly allow arbitrary nodes to be added to character units.

As there are no active games utilizing this functionality, I'm removing it as obsolete.