Vampire Hunter Boss Hack support
The Boss Hack ROM is misnamed, so we need to shift loads if and only if it's that version.  So I pushed the SFA2 byte sniffing code into shared gameclass code so that VH can use it.  Also moved the patch-specific code out to a new file since the gameclass file is fairly large.