palette/preview updates: blazblue, redearth, jojos; flag and fix bogus raws
The palette and preview updates speak for themselves as usual.

Some RAWs were using out-of-bound colors, in some cases intentionally and in some cases accidentally.  Updated imgdisp to note that to debugout, and updated imgout to be able to fix those RAWs on export if you turn that functionality on.