inverting blink; better ACT export/import; palette updates
Added an option for blink wherein we invert a color instead of just setting it to a selected color.
ACT export of secondary palettes was walking past the end of the palette as it presumed start=0 instead of start=secondary_palette_start.
ACT import for a node with multiple palettes could result in a secondary palette only receiving partial color assignment when a differently sized palette was applied.  Updated the code to only assign to a secondary palette if and only if we have complete coverage for that secondary palette.