PostSetPal: refresh onscreen paired palettes for updates
There's been a quirky design as regards paired palettes that are being shown onscreen when Update Palette runs through PostSetPal and updates one of those paired palettes.  Specifically for MvC2 Servbot, Cyclops, and Sentinel onscreen palettes are being changed but that has not been reflected in the UI.  This leads into slightly weird states wherein Undo would change the palette to be something not shown onscreen, as well as when you return to the palette that secondary palette is not what it was previously showing as.

This revised implementation updates visuals for the paired palettes to match what is in memory.  It's still a little quirky, but at least makes a more logical generalized path available.