modernize the game registration system
This consolidates the locations to plug in the game data that PalMod needs down to one single file (in addition to the new class for that game).  With this you just add the new game data to GameRegistry and the consuming code will obtain that data at runtime.

Also a few instances of code warning cleanup.