Fix drag and drop from Firefox
Firefox is down-converting all dragged images to a temporary palettized BMP.  That's actually better than Edge/Chrome which simply don't allow dragging images at all, but it is still useless.  However!  If we inspect the drop object we can get to the raw source image.  This change enables that pipeline: we'll create a temp file based upon the source image from the drop object, and use that if we can.