handle mouse-less slider usage
We bundle each strobe of mouse activity on the scroll bars into one Undo action.  However, that doesn't address mouse *wheel* modifications where we aren't getting MouseUp on completion.  So now we understandably treat ResetSlider as an end to all slider modifications against that slider instance.  That's a no-op for mouse pointer users but also handles the mouse wheel users.