Go back to old GIMP 2x RAW loading behavior
Per code note, it looks like there are two variants of funky GIMP RAW files.  Given that I currently now only have the "old" variety, switch to the code to use the "old" code.  If we find the new/other variety again, I'll make that work alongside the existing code that works for the "old" style.