custom previews: split multisprite compositions
When we import a PNG that represents multiple palettes being previewed we have the ability to decompose that PNG into the layers as defined by each palette.  This has to be a little lossy due to layering, but it does let us do somewhat interesting things such as using an older/newer Wolverine+Claws PNG preview for example.

This also removes "overflow" palette references on import.  So if you imported Wolverine+Claws (0x20 palette length) for Akuma, the imported preview would only use the 0x10 references, functionally splitting the import.

There's probably some interesting edge conditions to think further about, but this seems to cover most of the interesting applications well.