Enable auto-load of external previews
There's a long-term vision of treating Extras-based Unknown games as semi-equal partners.  This is part of that.  Here we allow you to use palette names as keys for loading external PNGs you have saved to a "previews" directory under PalMod.  These can be either RGB or Indexed PNGs: we'll handle using the same logic as drag-and-drop PNG preview logic.

There are two handlings for external Previews naming schemes.

1) For Extra palettes, we simply use "PALETTE_NAME.png".  If the palette contains what appears to be an index indicator such as "(2/200)" in "Foo (2/200)" we will truncate that and look for "Foo.png".  This allows us to cover multirange palette searches (consider an extra of "SearchMe \n 0 \n 0x20000") fairly effectively.
2) For internal palettes, we use two forks:
* If the node is not a named button color (A/B/C/LP/HK/etc), we use UNIT-NODE-PALETTE, as in "Akuma-Shared-Fireball.png".
* If the node is a named button color, we use UNIT-PALETTE, unless the PALETTE name itself includes UNIT and NODE.   So:
  "Gief | A | Cat"
becomes "Gief-Cat.png", and:
 "Gief | A | Gief A: Hammer"
gets logically reduced to Gief-Hammer.png, as needing to have both "Gief-Gief-A-Hammer.png" and "Gief-Gief-B-Hammer.png" seems dreadful.