improve export
We had some awkward and clunky code in place for export to palette files.  We were operating off the active palette, which can be non-intuitive to the user since the active palette is defined as the last palette interacted with.  Then we were operating on a page not palette basis.  As we have changed to have multiple possible palettes per page, walking per page is just wrong.  Modified the code to be solely operating off of loaded palettes, both ignoring page collections and the active palette.

Additionally, the export and reimport of GPL files was a little clunky given their lack of a zero color.  This was resulting in both a missing transparency byte on non-first palettes exported, and on import we would be shifting by 1 as we looped palettes. Handle that better for both export and import.