Preview import: add a common import/replacement dialog
When working on creating stage previews without the need of Photoshop, one aspect that was nettlesome was merging the preview of a palette that was shown on multiple graphic layers.  For example the background and foreground layer might both use the palette.  My solution to this is a new preview import dialog common to texture imports.  Its primary duty is to ensure we have correct dimensions for imported RAW files, but it also now handles directionality (allowing us to replace the previous Load Flipped Texture menu option), composition (enabling the merging scenario I needed), and layer selection.