ImgDisp: better handling of oddly sized 0th image in an image layer
If you have a 400x400 paired image and replace the 0th image with a 100x100 image, we ended up drawing the second image centered to 100x100, not 400x400.  This imperfect solution here ensures that the 0th image establishes centering on the currently largest known image.  If you inject 800x800 images for image 2 or such, it'll still look wonky, but this at least handles the scenario wherein you're starting to replace the stack.  Any further errors would be uninteresting and a continuation of pathological user behavior.