CPS3 madness! (+jojos map palettes)
As previously discussed, MAME is doing something interesting relative to what PalMod has been doing.  MAME and Capcom's Steam CPS3 both are using a RGB777 display space, not a RGB888 display space.  Given that Capcom is the gold standard, and that MAME is the other gold standard, update our logic to abide by their implementation.  The normal RGB555 code is there and we can potentially allow for that as an optional methodology if that is interesting.  But the important thing here is that now we can directly work with MAME/Steam sprites with no hoops needed.

Also added several Jojo's Story Mode palettes that were asked for.