Copy 15 colors option
This is the joy of writing your own code: you can do things that other people would consider just the quirkiest.  But bear with me, this is rational I swear.

The notion here is that if you are looking at a palette in an RGB image, importing all of that at once to PalMod is a pain in the butt.  CTRL+SHIFT+C (copy the current color under cursor), CTRL+SHIFT+V (copy that and paste it), and CTRL+SHIFT+N (copy that, paste it, and then move to the next palette index) are quite helpful, but it's still laborious.  This is quite imperfect code since we don't know source rect sizes - consider 00ff00 followed by 00ff00 - but it's potentially a huge time saver for copying palettes.

Copying 15 colors at once is not arbitrary: we don't have any ability to predict where the next set of colors would be in the source image.