ImgDump: don't needlessly suppress draws
We had some ancient broken code that I had fixed up from a no-op to a valid check... but it turns out that in that form it should have just been deleted.  The original broken code hinted at skipping needless resizes, but the check itself was never valid since the static was never updated.
 Once we started updating that static when you Exported the same size image back to back, we wouldn't display anything in the Preview until you changed the display size.

I maintain the idea in this checkin, but now the code is finally functional.