Code cleanup: imgdisp
As part of another potential project I ripped out a few clunky pieces of ImgDisp and made them their own classes.  As part of this I also updated our image storage for various paths.  It's still a little clunky and old in places, but this at least moves the ball forward.