ColorSystem: add tests, fix BRG444, wire in better alpha
PalMod is extensible, and as such there are color formats in it that we don't use and thus can't easily test.  Wrote a very simple test routine plugged in during debug runs.

This did catch an error in our unused BRG444 conversion that's now fixed.

Additionally, wired in support for None/Fixed/Variable alpha in for various color formats that didn't previously actively support those possibilities.