ImgDisp: don't autoload previews using mismatched sizes
If you have a custom preview filling in one (empty) spot for a preview set and that preview doesn't match the normal layout, the creaky old layout engine gets out of sorts.  This tweak avoids probably the only easy way to do that.

In theory the layout engine could use an overhaul, but it's functional for all normal usage.