Copy internals: show data per rom for multi-rom views
Our raw binary data view was a little cumbersome to work with when dealing with interleaved data sources.  This at least does the slicing by ROM for you and shows you that in the view as well.  Note that this doesn't work for the raw Copy Binary Data context menu option since I wouldn't know which ROM they wanted to look at.