MvC2: more cleanup, allow arbitrarily sized team views
Given the MST<->Roundhouse bug, I wanted to clean up the code further.  As part of that it's now easy enough to add support for 1...n size teams.  And given that I've had previous requests for some sort of "full" view... I added that crazy option.  Not sure how useful it is, but it's there now at the least.